Создал правильную файловую структуру, запилил namespas-ы, перенес скрипты в нужное место
