+++ Assets/MainScripts/Attack/EnemyAttack.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace MainScripts.Attack
+{
+    public class EnemyAttack: MonoBehaviour
+    {
+        public void Attack()
+        {
+            //Тут будем дергать метод HeroFallIntoTheFieldOfView() и если он вернет true тада атакуем
+        }
+    }
+}
