--- Assets/MainScripts/GameObject.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 585d8beb9b9d4db6bf57669e87dccbec
-timeCreated: 1680634383
