--- Assets/MainScripts/GameObject/IEnemyObjects.cs
@@ -1,16 +1,0 @@
-﻿using System.Numerics;
-
-namespace MainScripts.GameObject
-{
-    public interface IEnemyObjects
-    {
-        public void Attack();
-
-        public bool MoveByPointsAndMonitoring();
-
-        public bool HeroFallIntoTheFieldOfView();
-
-        public void RunToHeroAttack();
-        
-    }
-}
