--- Assets/MainScripts/GameObject/IEnemyObjects.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f9df40d4057443a8a508f9cc16911a1a
-timeCreated: 1680636666
