--- Assets/MainScripts/InfoHero.cs
@@ -1,12 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI;
-
-public class InfoHero : MonoBehaviour
-{
-    public float healthPoints = 74f;
-    public float maxHP = 100f;
-    public float damage = 25f;
-    public float attackRange = 1.417f;
-    public float armor = 5f;
-    public float speedAttack = 1f;
-}
