+++ Assets/MainScripts/Stats/HeroStats.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+namespace MainScripts.Stats
+{
+    public class HeroStats : MonoBehaviour
+    {
+        public float healthPoints = 74f;
+        public float maxHP = 100f;
+        public float damage = 25f;
+        public float attackRange = 1.417f;
+        public float armor = 5f;
+        public float speedAttack = 1f;
+    }
+}
