+++ Assets/Scenes/0 Scene - Village/Scripts/Dialogue.cs
@@ -1,63 +1,65 @@
 using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 using TMPro;
+using UnityEngine;
 
-public class Dialogue : MonoBehaviour
+namespace Scenes._0_Scene___Village.Scripts
 {
-    [SerializeField] TextMeshProUGUI text;
-    [SerializeField] string[] lines;
-    [SerializeField] float TextSpeed;
-
-    private int index;
-    // Start is called before the first frame update
-    void Start()
+    public class Dialogue : MonoBehaviour
     {
-        text.text = string.Empty;
-        StartDialogue();
+        [SerializeField] TextMeshProUGUI text;
+        [SerializeField] string[] lines;
+        [SerializeField] float TextSpeed;
 
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if(Input.GetMouseButtonDown(0))
+        private int index;
+        // Start is called before the first frame update
+        void Start()
         {
-            if(text.text == lines[index])
+            text.text = string.Empty;
+            StartDialogue();
+
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+            if(Input.GetMouseButtonDown(0))
             {
-                IsNextLine();
+                if(text.text == lines[index])
+                {
+                    IsNextLine();
+                }
+                else
+                {
+                    StopAllCoroutines();
+                    text.text = lines[index];
+                }
+            }
+        }
+
+        void StartDialogue(){
+            index = 0;
+            StartCoroutine(TypeLine());
+        }
+
+        IEnumerator TypeLine(){
+            foreach(char c in lines[index].ToCharArray())
+            {
+                text.text += c;
+                yield return new WaitForSeconds(TextSpeed);
+            }
+        }
+
+        void IsNextLine(){
+            if(index < lines.Length - 1)
+            {
+                index++;
+                text.text = string.Empty;
+                StartCoroutine(TypeLine());
             }
             else
             {
-                StopAllCoroutines();
-                text.text = lines[index];
+                gameObject.SetActive(false);
             }
-        }
-    }
-
-    void StartDialogue(){
-        index = 0;
-        StartCoroutine(TypeLine());
-    }
-
-    IEnumerator TypeLine(){
-        foreach(char c in lines[index].ToCharArray())
-        {
-            text.text += c;
-            yield return new WaitForSeconds(TextSpeed);
-        }
-    }
-
-    void IsNextLine(){
-        if(index < lines.Length - 1)
-        {
-            index++;
-            text.text = string.Empty;
-            StartCoroutine(TypeLine());
-        }
-        else
-        {
-            gameObject.SetActive(false);
         }
     }
 }
