+++ Assets/Scenes/0 Scene - Village/Scripts/Enemy/FireWizard.cs
@@ -1,0 +1,9 @@
+using MainScripts.Stats;
+
+namespace Scenes._0_Scene___Village.Scripts.Enemy
+{
+    public class FireWizard : EnemyStats
+    {
+    
+    }
+}
